% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/br_function.R
\name{br_function}
\alias{br_function}
\title{A BR Decider Function}
\usage{
br_function(type = "Chiller")
}
\arguments{
\item{type}{What kind of drink do you want? Defaults to Chiller.}
}
\description{
This function allows you to randomly pick a BR drink. Choose between a "Chiller" or "Smash Oatmeal"
}
\examples{
br_function()
}
\keyword{type}

